% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_param.R
\name{app_param}
\alias{app_param}
\title{init necessary parameter compilation to launch an app with authentification + account management}
\usage{
app_param(
  thumbnail_img = NULL,
  thumbnail_title = NULL,
  thumbnail_small_describe = NULL,
  categorie = NULL,
  modal_img_cover = NULL,
  modal_title = NULL,
  modal_description = NULL,
  modal_list_img = NULL
)
}
\arguments{
\item{thumbnail_img}{Path of the image in thumbnail.}

\item{thumbnail_title}{Title of the thumbnail.}

\item{thumbnail_small_describe}{Small description of the thumbnail.}

\item{categorie}{Categorie of the application.}

\item{modal_img_cover}{Image cover of the modal}

\item{modal_title}{Title of the application in the modal}

\item{modal_description}{Description of the application in the modal}

\item{modal_list_img}{List of small images of the application}
}
\value{
an error if provided information are incorrect a init function to launch in the app server part if success
}
\description{
init necessary parameter compilation to launch an app with authentification + account management
}
