% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_param.R
\name{app_param}
\alias{app_param}
\title{init necessary parameter compilation to launch an app with authentification + account management}
\usage{
app_param(
  thumbnail_img = NULL,
  thumbnail_title = NULL,
  thumbnail_small_describe = NULL,
  categorie = NULL,
  modal_img = NULL,
  modal_title = NULL
)
}
\arguments{
\item{thumbnail_img}{Path of the image in thumbnail}

\item{thumbnail_title}{Title of the thumbnail}

\item{modal_title}{}
}
\value{
an error if provided information are incorrect a init function to launch in the app server part if success
}
\description{
init necessary parameter compilation to launch an app with authentification + account management
}
